# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a test task for Effective Mobile",
        "title": "Test task Effective Mobile API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/songs": {
            "get": {
                "description": "Get a list of songs with filters and pagination",
                "summary": "Get songs with params",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group name",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Song name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Song id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/testEffectiveMobile_internal_models.Song"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.Failures"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a song with enrichment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create song with enrichment",
                "parameters": [
                    {
                        "description": "Name and group of the song",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controller.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.CreateSongResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.Failures"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.Failures"
                        }
                    }
                }
            }
        },
        "/songs/{id}": {
            "put": {
                "description": "Update song by id. You can provide partial updates, for example {\"text\": \"new text\"}.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update song by id (partial updates allowed)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Song id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Partial Song object",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.SongWithoutID"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "song updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.Failures"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.Failures"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete song by id",
                "summary": "Delete song by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Song id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "song deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.Failures"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.Failures"
                        }
                    }
                }
            }
        },
        "/songs/{id}/verses": {
            "get": {
                "description": "Get verses by song id with pagination",
                "summary": "Get verses by song id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Song id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "verses",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.Failures"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/testEffectiveMobile_internal_models.Failures"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_controller.Request": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "song": {
                    "type": "string"
                }
            }
        },
        "testEffectiveMobile_internal_models.CreateSongResponse": {
            "type": "object",
            "properties": {
                "song_id": {
                    "type": "integer"
                }
            }
        },
        "testEffectiveMobile_internal_models.Failures": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "testEffectiveMobile_internal_models.Song": {
            "description": "песня",
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "song": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "testEffectiveMobile_internal_models.SongWithoutID": {
            "type": "object",
            "properties": {
                "group": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "releaseDate": {
                    "type": "string"
                },
                "song": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        }
    }
}